requirements.txt
# Тестирпование
selenium==4.33.0
allure-pytest==2.15.2
allure-python-commons==2.15.2
pytest==9.0.1
pytest-xdist==3.8.0

# Форматтеры
black==25.1.0
isort==6.0.1